requirements.txt
streamlit 
requests
folium
streamlit_folium 
spacy
google.generativeai 
PyPDF2
torch